fix #260403: 自前レビューに従い、コンポーネントの分割→動作不良
