+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@tailwindcss/vite": "^4.2.2",
     "dnd-kit": "^0.0.2",
+    "lucide-react": "^1.7.0",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-markdown": "^10.1.0",
