+++ src/components/note-modal/NoteModalHeader.tsx
@@ -1,14 +1,18 @@
-interface ModalHeaderProps {
+import { X } from 'lucide-react';
+import type { Category } from '../../types';
+
+interface NoteHeaderProps {
   isEditing: boolean;
   editTitle: string;
   onTitleChange: (title: string) => void;
-  categoryLabel: string;
+  categoryLabel: Category;
   noteTitle: string;
   isPending: boolean;
   selectNote: (noteId: string | null) => void;
 }
 
-export const NoteModalHeader = ({ isEditing, editTitle, onTitleChange, categoryLabel, noteTitle, isPending, selectNote }: ModalHeaderProps) => {
+export const NoteHeader = ({ isEditing, editTitle, onTitleChange, categoryLabel, noteTitle, isPending, selectNote }: NoteHeaderProps) => {
+
   return (
     <div className="px-6 py-4 border-b flex justify-between items-center bg-gray-50">
       <div className="flex-1">
@@ -36,7 +40,7 @@ export const NoteModalHeader = ({ isEditing, editTitle, onTitleChange, categoryL
         onClick={() => selectNote(null)}
         className="ml-4 text-gray-400 hover:text-gray-600 transition-colors p-1"
       >
-        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><line x1="18" y1="6" x2="6" y2="18"></line><line x1="6" y1="6" x2="18" y2="18"></line></svg>
+        <X width="24" height="24" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round"><line x1="18" y1="6" x2="6" y2="18"></line><line x1="6" y1="6" x2="18" y2="18"></line></X>
       </button>
     </div>
   );
