+++ src/hooks/useNoteState.ts
@@ -6,16 +6,19 @@ import type { Category } from '../types';
 export const useNoteModalState = () => {
   const { notes, pendingNotes, selectNote } = useStore();
   const selectedNoteId = useStore((state) => state.selectedNoteId);
-  // ボード上のノートか、保留ボックスのノートかを探す
-  const note = notes.find((n) => n.id === selectedNoteId) ||
-    pendingNotes.find((n) => n.id === selectedNoteId);
 
   const [isEditing, setIsEditing] = useState(false);
   const [editTitle, setEditTitle] = useState('');
   const [editContent, setEditContent] = useState('');
   const [editCategory, setEditCategory] = useState<Category>('house');
+  const [newComment, setNewComment] = useState('');
+  const [isAppending, setIsAppending] = useState(false);
 
-  const isPending = note!.status === 'pending';
+  // ボード上のノートか、保留ボックスのノートかを探す
+  const note = notes.find((n) => n.id === selectedNoteId) ||
+    pendingNotes.find((n) => n.id === selectedNoteId);
+
+  const isPending = note?.status === 'pending';
 
   return {
     note,
@@ -27,6 +30,10 @@ export const useNoteModalState = () => {
     setEditContent,
     editCategory,
     setEditCategory,
+    newComment,
+    setNewComment,
+    isAppending,
+    setIsAppending,
     isPending,
     selectNote,
   };
