+++ src/utils/positionUtils.ts
