+++ 設計書.md
@@ -132,7 +132,7 @@ src/
 
 - **「折り畳み」と「詳細」の使い分け：**
 	- **通常時：** 付箋はタイトル（数文字）とカテゴリの色、投稿者アイコンのみを表示し、極力コンパクトに。
-    - **ホバー/タップ時：** 内容のプレビューがふわっと浮き出る。
+    - ~~**ホバー/タップ時：** 内容のプレビューがふわっと浮き出る。~~
     - **クリック時：** モーダル（またはサイドパネル）が開き、Markdownで詳細を編集・追記。    
 - **物理的な柔軟性：**
     - 4つの象限は固定ですが、境界線をドラッグして「今は『できること』が多いから、そのエリアを広げる」といった、**エリアサイズの可変性**を持たせると、1画面でも窮屈になりません。
